chore: Moved setLocationItems to the common folder
