+++ app/move/controllers/create/move-details.test.js
@@ -1,22 +1,12 @@
 const FormController = require('hmpo-form-wizard').Controller
 
-const referenceDataHelpers = require('../../../../common/helpers/reference-data')
+const commonMiddleware = require('../../../../common/middleware')
 const referenceDataService = require('../../../../common/services/reference-data')
 
 const BaseController = require('./base')
 const Controller = require('./move-details')
 
 const controller = new Controller({ route: '/' })
-const courtsMock = [
-  {
-    id: '8888',
-    title: 'Court 8888',
-  },
-  {
-    id: '9999',
-    title: 'Court 9999',
-  },
-]
 
 describe('Move controllers', function() {
   describe('Move Details', function() {
@@ -25,7 +15,7 @@ describe('Move controllers', function() {
         sinon.stub(BaseController.prototype, 'middlewareSetup')
         sinon.stub(controller, 'use')
         sinon.stub(controller, 'setMoveTypes')
-        sinon.stub(controller, 'setLocationItems')
+        sinon.stub(commonMiddleware, 'setLocationItems')
 
         controller.middlewareSetup()
       })
@@ -42,132 +32,18 @@ describe('Move controllers', function() {
 
       it('should call setLocationItems middleware', function() {
         expect(controller.use.secondCall).to.have.been.calledWith(
-          controller.setLocationItems()
+          commonMiddleware.setLocationItems()
         )
       })
 
       it('should call setLocationItems middleware', function() {
         expect(controller.use.thirdCall).to.have.been.calledWith(
-          controller.setLocationItems()
+          commonMiddleware.setLocationItems()
         )
       })
 
       it('should call correct number of middleware', function() {
         expect(controller.use.callCount).to.equal(3)
-      })
-    })
-
-    describe('#setLocationItems()', function() {
-      let req, res, nextSpy
-
-      beforeEach(function() {
-        req = {
-          form: {
-            options: {
-              fields: {
-                to_location_court: {},
-              },
-            },
-          },
-        }
-        res = {}
-        nextSpy = sinon.spy()
-
-        sinon.stub(referenceDataHelpers, 'filterDisabled').callsFake(() => {
-          return () => true
-        })
-        sinon.stub(referenceDataService, 'getLocationsByType')
-      })
-
-      context('when field exists', function() {
-        const mockFieldName = 'to_location_court'
-        const mockLocationType = 'court'
-
-        context('when service resolves', function() {
-          beforeEach(async function() {
-            referenceDataService.getLocationsByType.resolves(courtsMock)
-
-            await controller.setLocationItems(mockLocationType, mockFieldName)(
-              req,
-              {},
-              nextSpy
-            )
-          })
-
-          it('should call reference data service', function() {
-            expect(
-              referenceDataService.getLocationsByType
-            ).to.be.calledOnceWithExactly(mockLocationType)
-          })
-
-          it('populates the move type items', function() {
-            expect(req.form.options.fields[mockFieldName].items).to.deep.equal([
-              {
-                text: `--- Choose ${mockLocationType} ---`,
-              },
-              {
-                text: 'Court 8888',
-                value: '8888',
-              },
-              {
-                text: 'Court 9999',
-                value: '9999',
-              },
-            ])
-          })
-
-          it('should call next', function() {
-            expect(nextSpy).to.be.calledOnceWithExactly()
-          })
-        })
-
-        context('when service rejects', function() {
-          const errorMock = new Error('Problem')
-
-          beforeEach(async function() {
-            referenceDataService.getLocationsByType.throws(errorMock)
-
-            await controller.setLocationItems(mockLocationType, mockFieldName)(
-              req,
-              {},
-              nextSpy
-            )
-          })
-
-          it('should call next with the error', function() {
-            expect(nextSpy).to.be.calledOnceWithExactly(errorMock)
-          })
-
-          it('should not mutate request object', function() {
-            expect(req).to.deep.equal({
-              form: {
-                options: {
-                  fields: {
-                    to_location_court: {},
-                  },
-                },
-              },
-            })
-          })
-        })
-      })
-
-      context('when field does not exist', function() {
-        beforeEach(async function() {
-          await controller.setLocationItems('court', 'non_existent')(
-            req,
-            res,
-            nextSpy
-          )
-        })
-
-        it('should not call reference service', function() {
-          expect(referenceDataService.getLocationsByType).not.to.be.called
-        })
-
-        it('should call next', function() {
-          expect(nextSpy).to.be.calledOnceWithExactly()
-        })
       })
     })
