+++ common/middleware/index.js
@@ -1,7 +1,9 @@
 const setDatePeriod = require('./set-date-period')
 const setDateRange = require('./set-date-range')
+const setLocationItems = require('./set-location-items')
 
 module.exports = {
   setDatePeriod,
   setDateRange,
+  setLocationItems,
 }
