+++ common/middleware/set-location-items.js
@@ -1,0 +1,33 @@
+const { set } = require('lodash')
+
+const fieldHelpers = require('../helpers/field')
+const referenceDataHelpers = require('../helpers/reference-data')
+const referenceDataService = require('../services/reference-data')
+
+function setLocationItems(locationType, fieldName) {
+  return async (req, res, next) => {
+    const { fields } = req.form.options
+
+    if (!fields[fieldName]) {
+      return next()
+    }
+
+    try {
+      const locations = await referenceDataService.getLocationsByType(
+        locationType
+      )
+      const items = fieldHelpers.insertInitialOption(
+        locations
+          .filter(referenceDataHelpers.filterDisabled())
+          .map(fieldHelpers.mapReferenceDataToOption),
+        locationType
+      )
+
+      set(req, `form.options.fields.${fieldName}.items`, items)
+      next()
+    } catch (error) {
+      next(error)
+    }
+  }
+}
+module.exports = setLocationItems
